move utils functions into alerts/utils
