+++ static/app/views/alerts/utils/constants.tsx
